Maybe type is now an opt-in extension
