+++ lib/dry/types/extensions.rb
@@ -1,0 +1,3 @@
+Dry::Types.register_extension(:maybe) do
+  require 'dry/types/extensions/maybe'
+end
