+++ spec/dry/types/default_spec.rb
@@ -1,4 +1,8 @@
 RSpec.describe Dry::Types::Definition, '#default' do
+  before(:all) do
+    Dry::Types.load_extensions(:maybe)
+  end
+
   context 'with a definition' do
     subject(:type) { Dry::Types['string'].default('foo') }
