+++ spec/dry/types/maybe_spec.rb
@@ -1,4 +1,8 @@
 RSpec.describe Dry::Types::Definition, '#maybe' do
+  before(:all) do
+    Dry::Types.load_extensions(:maybe)
+  end
+
   context 'with a definition' do
     subject(:type) { Dry::Types['string'].maybe }
