+++ spec/spec_helper.rb
@@ -45,4 +45,8 @@ RSpec.configure do |config|
 
     Object.send(:remove_const, Test.remove_constants.name)
   end
+
+  config.order = 'random'
 end
+
+srand RSpec.configuration.seed
